+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## 3.0.0 (2020-01-13)
 
 - Breaking change: Expect Node >= 8
+- Fix: Require `json` as "own" property
 - Fix: wrap: false returning inconsistent data types (@CacheControl)
 - Fix: Ensure throwing with a bad result type
 - Fix: Avoid erring when value before parent selector is falsey